create build to compute report of detMorse
git-svn-id: file:///vagrant/svn/dsis@891 6a918f4f-bbb9-b144-964e-14738c76912e

+++ trunk/dsis/src/Graphs/Graph.Morse/src/ComputationResult.cs
@@ -5,9 +5,9 @@ namespace DSIS.Graph.Morse
   public class ComputationResult
   {
     public readonly double Value;
-    public readonly double Count;
+    public readonly int Count;
 
-    public ComputationResult(double value, double count)
+    public ComputationResult(double value, int count)
     {
       Value = value;
       Count = count;
